# boot/assets/sw/config/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is builtin RK common service.",
        "title": "RK Common Service",
        "contact": {
            "name": "rk-dev",
            "url": "https://github.com/rookie-ninja/rk-gf",
            "email": "lark@pointgoal.io"
        },
        "license": {
            "name": "Apache 2.0 License",
            "url": "https://github.com/rookie-ninja/rk-gf/blob/master/LICENSE.txt"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/rk/v1/tv/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "text/html"
                ],
                "summary": "Get HTML page of /tv/",
                "operationId": "15",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}